clean BuildFiles up to before DataFormats
