+++ AnalysisAlgos/TrackInfoProducer/test/BuildFile.xml
@@ -1,17 +1,11 @@
-<use name="FWCore/Utilities"/>
 <use name="FWCore/Framework"/>
 <use name="AnalysisDataFormats/TrackInfo"/>
 <use name="DataFormats/SiStripDetId"/>
 <use name="DataFormats/TrackingRecHit"/>
 <use name="DataFormats/TrackerRecHit2D"/>
 <use name="DataFormats/TrackReco"/>
-<use name="DataFormats/Common"/>
-<use name="Geometry/CommonDetUnit"/>
 <use name="Geometry/Records"/>
-<use name="Geometry/TrackerGeometryBuilder"/>
 <use name="clhep"/>
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/ParameterSet"/>
 <use name="root"/>
 <library file="TrackInfoAnalyzer.cc" name="TrackInfoAnalyzer">
   <flags EDM_PLUGIN="1"/>
