+++ AnalysisDataFormats/EWK/BuildFile.xml
@@ -1,12 +1,7 @@
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/ParameterSet"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/Candidate"/>
-<use name="DataFormats/JetReco"/>
 <use name="DataFormats/METReco"/>
 <use name="CommonTools/CandUtils"/>
-<use name="CommonTools/UtilAlgos"/>
-<use name="CommonTools/RecoAlgos"/>
 <use name="CLHEP"/>
 <export>
   <lib name="1"/>
