+++ AnalysisDataFormats/TopObjects/BuildFile.xml
@@ -1,13 +1,7 @@
 <use name="FWCore/Utilities"/>
 <use name="DataFormats/Common"/>
-<use name="DataFormats/TauReco"/>
-<use name="DataFormats/JetReco"/>
-<use name="DataFormats/METReco"/>
-<use name="DataFormats/MuonReco"/>
-<use name="DataFormats/TrackReco"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/HepMCCandidate"/>
-<use name="DataFormats/EgammaCandidates"/>
 <use name="CommonTools/CandUtils"/>
 <use name="root"/>
 <export>
