+++ AnalysisDataFormats/TrackInfo/BuildFile.xml
@@ -4,9 +4,7 @@
 <use name="DataFormats/TrackingRecHit"/>
 <use name="DataFormats/TrajectorySeed"/>
 <use name="DataFormats/TrajectoryState"/>
-<use name="SimDataFormats/GeneratorProducts"/>
 <use name="SimDataFormats/TrackingAnalysis"/>
-<use name="TrackingTools/PatternTools"/>
 <use name="FWCore/MessageLogger"/>
 <export>
   <lib name="1"/>
