+++ CalibTracker/SiStripCommon/plugins/SealModules.cc
@@ -8,7 +8,6 @@ DEFINE_FWK_MODULE(SiStripDetInfoFileWriter);
 #include "CalibTracker/SiStripCommon/interface/ShallowTree.h"
 #include "CalibTracker/SiStripCommon/interface/ShallowEventDataProducer.h"
 #include "CalibTracker/SiStripCommon/interface/ShallowDigisProducer.h"
-#include "CalibTracker/SiStripCommon/interface/ShallowClustersProducer.h"
 #include "CalibTracker/SiStripCommon/interface/ShallowTrackClustersProducer.h"
 #include "CalibTracker/SiStripCommon/interface/ShallowRechitClustersProducer.h"
 #include "CalibTracker/SiStripCommon/interface/ShallowSimhitClustersProducer.h"
@@ -19,7 +18,6 @@ DEFINE_FWK_MODULE(SiStripDetInfoFileWriter);
 DEFINE_FWK_MODULE(ShallowTree);
 DEFINE_FWK_MODULE(ShallowEventDataProducer);
 DEFINE_FWK_MODULE(ShallowDigisProducer);
-DEFINE_FWK_MODULE(ShallowClustersProducer);
 DEFINE_FWK_MODULE(ShallowTrackClustersProducer);
 DEFINE_FWK_MODULE(ShallowRechitClustersProducer);
 DEFINE_FWK_MODULE(ShallowSimhitClustersProducer);
