+++ CaloOnlineTools/HcalOnlineDb/plugins/BuildFile.xml
@@ -1,15 +1,9 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/Utilities"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
 <use name="boost"/>
-<use name="OnlineDB/Oracle"/>
-<use name="DataFormats/Common"/>
 <use name="xerces-c"/>
 <use name="zlib"/>
 <use name="DataFormats/HcalDetId"/>
-<use name="CalibCalorimetry/CaloTPG"/>
 <use name="CalibCalorimetry/HcalTPGAlgos"/>
 <use name="CaloOnlineTools/HcalOnlineDb"/>
 <library file="*.cc" name="CaloOnlineToolsHcalOnlineDbPlugins">
