+++ CommonTools/Clustering1D/test/BuildFile.xml
@@ -1,4 +1,3 @@
 <use name="CommonTools/Clustering1D"/>
-<use name="FWCore/Framework"/>
 <bin file="ClusteringTest.cpp">
 </bin>
