+++ CommonTools/MVAUtils/plugins/BuildFile.xml
@@ -1,8 +1,6 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="CondCore/DBOutputService"/>
-<use name="CondFormats/EgammaObjects"/>
 <use name="CommonTools/MVAUtils"/>
 <use name="boost"/>
 <flags EDM_PLUGIN="1"/>
