+++ CommonTools/PileupAlgos/BuildFile.xml
@@ -1,5 +1,3 @@
-<use name="DataFormats/Common"/>
-<use name="DataFormats/ParticleFlowCandidate"/>
 <use name="FWCore/ParameterSet"/>
 <use name="root"/>
 <use name="rootcore"/>
