+++ CommonTools/Utils/plugins/BuildFile.xml
@@ -1,4 +1,3 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <flags EDM_PLUGIN="1"/>
