+++ CondCore/PopCon/BuildFile.xml
@@ -1,6 +1,7 @@
+<use name="CondCore/CondDB"/>
+<use name="CondCore/DBOutputService"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
-<use name="CondCore/DBOutputService"/>
 <export>
   <lib name="1"/>
 </export>
