+++ CondCore/SiPixelPlugins/BuildFile.xml
@@ -1,4 +1,5 @@
 <use name="CalibTracker/StandaloneTrackerTopology"/>
+<use name="CalibTracker/SiPixelESProducers"/>
 <use name="CondCore/ESSources"/>
 <use name="CondFormats/SiPixelObjects"/>
 <use name="CondFormats/DataRecord"/>
