+++ CondTools/DQM/BuildFile.xml
@@ -1,4 +1,4 @@
-<use name="CondCore/CondDB"/>
+<use name="CondCore/PopCon"/>
 <use name="CondFormats/Common"/>
 <use name="CondFormats/DQMObjects"/>
 <use name="CoralBase"/>
