+++ Configuration/Skimming/BuildFile.xml
@@ -8,20 +8,14 @@
 <use name="DataFormats/JetReco"/>
 <use name="DataFormats/Common"/>
 <use name="Geometry/CaloTopology"/>
-<use name="Geometry/Records"/>
 <use name="RecoEcal/EgammaCoreTools"/>
-<use name="RecoJets/JetAlgorithms"/>
 <use name="HLTrigger/HLTcore"/>
-<use name="CommonTools/UtilAlgos"/>
-<use name="PhysicsTools/UtilAlgos"/>
 <use name="DataFormats/Candidate"/>
-<use name="DataFormats/Luminosity"/>
 <use name="DataFormats/TrackReco"/>
 <use name="RecoVertex/KalmanVertexFit"/>
 <use name="TrackingTools/IPTools"/>
 <use name="MagneticField/Engine"/>
 <use name="MagneticField/ParametrizedEngine"/>
-<use name="TrackingTools/PatternTools"/>
 <use name="TrackingTools/TransientTrack"/>
 <use name="TrackingTools/Records"/>
 <flags EDM_PLUGIN="1"/>
