+++ DPGAnalysis/HcalTools/BuildFile.xml
@@ -3,10 +3,7 @@
 <use name="DataFormats/HcalRecHit"/>
 <use name="EventFilter/HcalRawToDigi"/>
 <use name="Geometry/CaloGeometry"/>
-<use name="DataFormats/GeometryVector"/>
 <use name="Geometry/HcalCommonData"/>
-<use name="DataFormats/TrackingRecHit"/>
-<use name="CommonTools/UtilAlgos"/>
 <use name="DataFormats/FEDRawData"/>
 <use name="CalibFormats/HcalObjects"/>
 <use name="CalibCalorimetry/HcalAlgos"/>
