+++ DPGAnalysis/SiStripTools/plugins/BuildFile.xml
@@ -1,7 +1,6 @@
 <use name="root"/>
 <use name="boost"/>
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ServiceRegistry"/>
@@ -22,13 +21,11 @@
 <use name="DataFormats/TCDS"/>
 <use name="DPGAnalysis/SiStripTools"/>
 <use name="DQM/SiStripCommon"/>
-<use name="DQMServices/Core"/>
 <use name="CommonTools/TrackerMap"/>
 <use name="CalibFormats/SiStripObjects"/>
 <use name="CondFormats/SiPixelObjects"/>
 <use name="CondFormats/DataRecord"/>
 <use name="CalibTracker/Records"/>
-<use name="CalibTracker/SiStripCommon"/>
 <use name="CalibTracker/SiPixelESProducers"/>
 <use name="MagneticField/Records"/>
 <use name="MagneticField/Engine"/>
@@ -52,8 +49,6 @@
 <use name="SimDataFormats/TrackingAnalysis"/>
 <use name="SimDataFormats/TrackingHit"/>
 <use name="DataFormats/TrackingRecHit"/>
-<use name="SimTracker/TrackAssociation"/>
-<use name="SimTracker/Records"/>
 <library file="*.cc" name="DPGAnalysisSiStripToolsPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
