--- DQM/SiOuterTracker/BuildFile.xml
@@ -1,7 +1,0 @@
-<use name="DataFormats/Common"/>
-<use name="DataFormats/L1TrackTrigger"/>
-<use name="DQMServices/Core"/>
-<use name="FWCore/Utilities"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/ParameterSet"/>
-<flags EDM_PLUGIN="1"/>
