+++ DQMServices/Core/BuildFile.xml
@@ -2,7 +2,6 @@
 <use name="FWCore/ParameterSet"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
-<use name="FWCore/Version"/>
 <use name="classlib"/>
 <use name="roothistmatrix"/>
 <use name="protobuf"/>
