+++ EgammaAnalysis/ElectronTools/BuildFile.xml
@@ -1,23 +1,16 @@
 <use name="CommonTools/Egamma"/>
-<use name="FWCore/Framework"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
-<use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
 <use name="Geometry/CaloTopology"/>
-<use name="CondCore/DBOutputService"/>
 <use name="CondFormats/EgammaObjects"/>
-<use name="CondFormats/PhysicsToolsObjects"/>
 <use name="RecoEcal/EgammaCoreTools"/>
-<use name="DataFormats/Candidate"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/EgammaCandidates"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/EgammaReco"/>
 <use name="DataFormats/TrackReco"/>
 <use name="TrackingTools/IPTools"/>
-<use name="PhysicsTools/UtilAlgos"/>
-<use name="PhysicsTools/SelectorUtils"/>
 <use name="RecoEgamma/EgammaTools"/>
 <use name="clhep"/>
 <use name="root"/>
