+++ EgammaAnalysis/ElectronTools/plugins/BuildFile.xml
@@ -1,16 +1,12 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/VertexReco"/>
-<use name="PhysicsTools/UtilAlgos"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="Geometry/CaloTopology"/>
 <use name="Geometry/Records"/>
 <use name="EgammaAnalysis/ElectronTools"/>
-<use name="RecoEcal/EgammaCoreTools"/>
-<use name="PhysicsTools/SelectorUtils"/>
 <use name="TrackingTools/Records"/>
 <library name="EgammaAnalysisElectronToolsPlugins" file="*.cc">
   <flags EDM_PLUGIN="1"/>
