+++ ElectroWeakAnalysis/Utilities/BuildFile.xml
@@ -1,13 +1,10 @@
-<use name="CommonTools/Utils"/>
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/Candidate"/>
-<use name="DataFormats/PatCandidates"/>
-<use name="CommonTools/UtilAlgos"/>
+<use name="DataFormats/METReco"/>
 <use name="CommonTools/CandUtils"/>
-<use name="RecoMuon/Records"/>
+<use name="CondFormats/DataRecord"/>
 <use name="CondFormats/RecoMuonObjects"/>
 <use name="MuonAnalysis/MomentumScaleCalibration"/>
 <use name="lhapdf"/>
