+++ ElectroWeakAnalysis/WENu/BuildFile.xml
@@ -1,9 +1,7 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/Common"/>
-<use name="RecoLocalCalo/EcalRecAlgos"/>
 <use name="root"/>
 <use name="rootmath"/>
 <use name="rootcore"/>
