+++ ElectroWeakAnalysis/WMuNu/BuildFile.xml
@@ -1,13 +1,9 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/Common"/>
-<use name="DataFormats/Candidate"/>
 <use name="DataFormats/JetReco"/>
 <use name="DataFormats/METReco"/>
 <use name="AnalysisDataFormats/EWK"/>
-<use name="CommonTools/CandUtils"/>
 <use name="CommonTools/UtilAlgos"/>
-<use name="CommonTools/RecoAlgos"/>
 <use name="CLHEP"/>
 <flags EDM_PLUGIN="1"/>
