+++ ElectroWeakAnalysis/WMuNu/bin/BuildFile.xml
@@ -1,3 +1,3 @@
-<use name="PhysicsTools/Utilities"/>
 <use name="rootrint"/>
+<use name="rootgraphics"/>
 <bin name="WMuNuValidatorMacro" file="WMuNuValidatorMacro.cpp"/>
