+++ ElectroWeakAnalysis/ZEE/BuildFile.xml
@@ -1,14 +1,11 @@
 <use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
 <use name="FWCore/ParameterSet"/>
 <use name="DataFormats/PatCandidates"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/Scalers"/>
 <use name="Geometry/CaloGeometry"/>
 <use name="Geometry/CaloTopology"/>
-<use name="CommonTools/Egamma"/>
 <use name="CommonTools/UtilAlgos"/>
-<use name="L1Trigger/GlobalTriggerAnalyzer"/>
 <use name="MagneticField/Records"/>
 <use name="MagneticField/Engine"/>
 <use name="RecoEcal/EgammaCoreTools"/>
